pyproject.toml
[tool.poetry]
name = "geodjango-starter"
version = "3.0.2"
description = "Start GeoDjango easily. [Django, REST framework GIS, PostgreSQL, PostGIS]"
authors = ["Yasunori Kirimoto"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.2"
djangorestframework-gis = "^0.15"
django-filter = "^2.2.0"
psycopg2-binary = "^2.8.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
